✨api 연동 CRUD

+++ react_study/src/app/Control.tsx
@@ -1,37 +1,64 @@
 "use client";
 import Link from "next/link";
 import { useParams, useRouter } from "next/navigation";
+import { useState } from "react";
 
 export function Control() {
   const router = useRouter();
   const params = useParams();
-  const post_id = params.post_id;
+  const id = params.id;
+
+  const [password, setPassword] = useState("");
+
+  const handleDelete = async () => {
+    if (!password) {
+      alert("비밀번호를 입력하세요.");
+      return;
+    }
+
+    try {
+      const resp = await fetch(`${process.env.NEXT_PUBLIC_API_URL}posts`, {
+        method: 'DELETE',
+        headers: {
+          'Content-Type': 'application/json', 
+        },
+        body: JSON.stringify({
+          post_id: id,      
+          user_password: password, 
+        }),
+      });
+
+      if (!resp.ok) {
+        throw new Error(`Failed to delete topic with ID: ${id}`);
+      }
+
+      router.push('/');
+      router.refresh();
+      console.log("삭제 성공");
+    } catch (error) {
+      console.error("Delete failed:", error);
+      alert("주제 삭제에 실패했습니다. 다시 시도해 주세요.");
+    }
+  };
 
   return (
     <ul>
       <li>
         <Link href="/create">create</Link>
       </li>
-      {post_id ? (
+      {id ? (
         <>
           <li>
-            <Link href={`/update/${post_id}`}>update</Link>
+            <Link href={`/update/${id}`}>update</Link>
           </li>
           <li>
-            <button
-              onClick={async () => {
-                const resp = await fetch(`http://api.ddubbu.com:8000/posts/${post_id}`, {
-                  method: "DELETE",
-                });
-
-                if (resp.ok) {
-                  router.push("/");
-                  router.refresh();
-                } else {
-                  console.error("Failed to delete the post");
-                }
-              }}
-            >
+            <input
+              type="password"
+              placeholder="비밀번호 입력"
+              value={password}
+              onChange={(e) => setPassword(e.target.value)}
+            />
+            <button onClick={handleDelete}>
               delete
             </button>
           </li>
